www.apply: move RefuseViewStep to batch_views
